bookingsiden app.js er rykket til app.booking.js i views for at opfylde SPA. Den virker og den gamle er slettet. Kun eventlistener er slettet fra den oprindelige, da jeg har lavet en routing JS i stedeet Denne håndterer alle nagivationer og hedder main.js
